refactor: signup login page component separation
